requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.3
django-environ==0.10.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
filelock==3.12.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
pbr==5.11.1
Pillow==10.0.0
platformdirs==3.8.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==1.5
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.7.1
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==6.5.0
